An approach to organize configuration better
