+++ src/main/java/de/bywambo/chattodiscord/config/ResponseCodes.java
@@ -1,13 +1,15 @@
-package de.bywambo.chattodiscord.statics;
+package de.bywambo.chattodiscord.config;
 
-public class ResponseCodes {
+/**
+ * Important Discord status codes from
+ * <a href="https://discord.com/developers/docs/topics/opcodes-and-status-codes#http">discord.com</a>
+ */
+public final class ResponseCodes {
 
     private ResponseCodes() {
         throw new IllegalStateException("Utility class");
     }
 
-    // Important Discord status codes from
-    // https://discord.com/developers/docs/topics/opcodes-and-status-codes#http
     public static final int OK_NO_CONTENT = 204;
     public static final int BAD_REQUEST = 400;
     public static final int UNAUTHORIZED = 401;
