--- src/main/java/de/bywambo/chattodiscord/utils/Config.java
@@ -1,14 +1,0 @@
-package de.bywambo.chattodiscord.utils;
-
-import java.net.URL;
-
-public class Config {
-
-    private Config() {
-        throw new IllegalStateException("Utility class");
-    }
-
-    public static boolean Enabled = false;
-    public static URL WebhookURL = null;
-
-}
