+++ src/main/java/de/bywambo/chattodiscord/utils/Connector.java
@@ -1,6 +1,6 @@
 package de.bywambo.chattodiscord.utils;
 
-import de.bywambo.chattodiscord.statics.ResponseCodes;
+import de.bywambo.chattodiscord.config.ResponseCodes;
 import org.bukkit.Bukkit;
 
 import java.io.IOException;
